Split CLI into multiple files, improve help
